Create battle_of_names.py
You will receive a number N. On the following N lines, you will be receiving names. You should sum the ASCII values of each letter in the name and integer divide it by the number of the current row (starting from 1). 

Save the result to a set of either odd or even numbers, depending on if the resulting number is odd or even. After that, sum the values of each set.

If the sums of the two sets are equal, print the union of the values, separated by ", ". 

If the sum of the odd numbers is bigger than the sum of the even numbers, print the different values, separated by ", ".

If the sum of the even numbers is bigger than the sum of the odd numbers, print the symmetric-different values, separated by ", ".